+++ crates/cojson-core-napi/npm/linux-arm64-gnu/CHANGELOG.md
@@ -1,5 +1,7 @@
 # cojson-core-napi-linux-arm64-gnu
 
+## 0.18.32
+
 ## 0.18.31
 
 ## 0.18.30